+++ app/src/main/AndroidManifest.xml
@@ -13,11 +13,8 @@
         android:theme="@style/AppTheme">
         <activity android:name=".ToDoListActivity"></activity>
         <activity android:name=".WeatherActivity" />
-        <activity android:name=".ViewAnimalDetailsActivity" />
         <activity android:name=".UpdateAnimalActivity" />
         <activity android:name=".AddAnimalActivity" />
-        <activity android:name=".EmployeeActivity" />
-        <activity android:name=".ExpenditureActivity" />
         <activity android:name=".VetActivity" />
         <activity android:name=".AnimalActivity" />
         <activity android:name=".LoginActivity" />
