--- app/src/main/java/com/example/farm/EmployeeActivity.java
@@ -1,41 +1,0 @@
-package com.example.farm;
-
-import androidx.annotation.NonNull;
-import androidx.appcompat.app.AppCompatActivity;
-
-import android.content.Intent;
-import android.os.Bundle;
-import android.view.Menu;
-import android.view.MenuInflater;
-import android.view.MenuItem;
-import android.widget.Toast;
-
-public class EmployeeActivity extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_employee);
-        Intent intent = getIntent();
-
-    }
-
-    public boolean onCreateOptionsMenu(Menu menu){
-        MenuInflater inflater = getMenuInflater();
-        inflater.inflate(R.menu.menumainopts, menu);
-        return true;
-    }
-
-    @Override
-    public boolean onOptionsItemSelected(@NonNull MenuItem item) {
-        switch (item.getItemId()){
-            case R.id.menuLogout:
-                Toast.makeText(this, "Logging out", Toast.LENGTH_SHORT).show();
-                startActivity(new Intent(EmployeeActivity.this, LoginActivity.class));
-            case R.id.menuHome:
-                Toast.makeText(this, "Home Page", Toast.LENGTH_SHORT).show();
-                startActivity(new Intent(EmployeeActivity.this, MainActivity.class));
-        }
-        return super.onOptionsItemSelected(item);
-    }
-}
