+++ app/src/main/java/com/example/farm/ToDoListActivity.java
@@ -14,6 +14,9 @@ import android.content.Intent;
 import android.graphics.Canvas;
 import android.os.Bundle;
 import android.util.Log;
+import android.view.Menu;
+import android.view.MenuInflater;
+import android.view.MenuItem;
 import android.view.View;
 import android.widget.EditText;
 import android.widget.Toast;
@@ -259,5 +262,26 @@ public class ToDoListActivity extends AppCompatActivity implements FirebaseAuth.
             toDoListAdapter.stopListening();
         }
     }
+
+    public boolean onCreateOptionsMenu(Menu menu){
+        MenuInflater inflater = getMenuInflater();
+        inflater.inflate(R.menu.menumainopts, menu);
+        return true;
+
+    }
+
+    @Override
+    public boolean onOptionsItemSelected(@NonNull MenuItem item) {
+        switch (item.getItemId()){
+            case R.id.menuLogout:
+                Toast.makeText(this, "Logging out", Toast.LENGTH_SHORT).show();
+                startActivity(new Intent(ToDoListActivity.this, LoginActivity.class));
+            case R.id.menuHome:
+                Toast.makeText(this, "Home Page", Toast.LENGTH_SHORT).show();
+                startActivity(new Intent(ToDoListActivity.this, MainActivity.class));
+        }
+        return super.onOptionsItemSelected(item);
+    }
+
 
 }
