--- app/src/main/java/com/example/farm/ViewAnimalDetailsActivity.java
@@ -1,14 +1,0 @@
-package com.example.farm;
-
-import androidx.appcompat.app.AppCompatActivity;
-
-import android.os.Bundle;
-
-public class ViewAnimalDetailsActivity extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_view_animal_details);
-    }
-}
