+++ app/src/main/res/layout/activity_login.xml
@@ -4,6 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
+    android:background="@drawable/grassland"
     tools:context=".LoginActivity">
 
     <EditText
