+++ app/src/main/res/values/colors.xml
@@ -6,4 +6,5 @@
     <color name="colorRed">#F10606</color>
     <color name="white">#FFFFFF</color>
     <color name="colorBlack">#FF090101</color>
+    <color name="colorSky">#95E2DE</color>
 </resources>
